+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
-### Unreleased
+### v2.0.152 22 June 2023
+
 - [new] Add ability to change save location on mac
 - [new] Improve file location dialog ui
 - [new] Add an experimental option to run the app in the menu bar on macos
@@ -9,11 +10,13 @@
 - [fix] Remove receive status after url received
 
 ### v2.0.146 19 November 2022
+
 - [new] Enable sending of multiple files
 - [new] Add ability to change save location on windows and linux
 - [new] Add option to open received files folder on windows, mac and linux
 
 ### v2.0.142 15 November 2022
+
 - [new] Linux support!
 - [improvement] Improve app icon on Android
 - [internal] Dependency upgrades
@@ -21,10 +24,12 @@
 - [internal] Remove custom version script for msix package
 
 ### v2.0.141 14 November 2022
+
 - [internal] Improved error and analytics logs
 - [internal] Upgrade to flutter 3.3
 
 ### v2.0.134 15 August 2022
+
 - [internal] Use strong mode for implicit-casts and implicit-dynamic
 - [internal] Use simple_peer library for connection
 - [internal] Update third party libraries
@@ -32,10 +37,12 @@
 - [internal] Use higher timeout for data receiver
 
 ### v2.0.131 27 July 2022
+
 - [new] Support for Apple Silicon
 - [improvement] Set minimum window size on windows
 - [fix] Open files app instead of gallery for file selection
 - [internal] Update third party libraries
 
 ### v2.0.130 23 July 2022
+
 - [internal] Mostly refactoring and fixes in preparation for open source